TCF Agent: macro defs _FILE_OFFSET_BITS=64 _LARGEFILE_SOURCE _GNU_SOURCE are moved to .c files that need them instead of being global.

+++ framework/channel.c
@@ -76,14 +76,14 @@ static void write_block_all(OutputStream * out, const char * bytes, size_t size)
     }
 }
 
-static ssize_t splice_block_all(OutputStream * out, int fd, size_t size, off_t * offset) {
+static ssize_t splice_block_all(OutputStream * out, int fd, size_t size, int64_t * offset) {
     char buffer[0x400];
     ssize_t rd = 0;
 
     assert(is_dispatch_thread());
     if (size > sizeof(buffer)) size = sizeof(buffer);
     if (offset != NULL) {
-        rd = pread(fd, buffer, size, *offset);
+        rd = pread(fd, buffer, size, (off_t)*offset);
         if (rd > 0) *offset += rd;
     }
     else {
